requirements.txt
opencv-python==4.8.1.78
setuptools
numpy<2
openai-whisper
pyaudio
